breederCMS/requirements.txt
Django==1.11.2
django-appconf==1.0.1
django-daguerre==2.2.3
olefile==0.44
Pillow==4.1.1
pytz==2017.2
six==1.10.0
